requirements/defaults.txt
Django==1.8.4
Fabric==1.10.1
Pillow==2.9.0
amqp==1.4.6
anyjson==0.3.3
astropy==1.0.4
beautifulsoup4==4.4.0
billiard==3.3.0.20
celery==3.1.18
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.23.0
django-autoslug==1.7.2
django-bower==5.0.4
django-contact-form==1.1
django_compressor==1.5
django-cors-headers==1.1.0
django-hosts==1.2
django-honeypot==0.5.0
django.js==0.8.1
django-lockdown==1.1
django-multiselectfield==0.1.3
django-nose==1.4.0
django-meta==0.3.1
django-settings-export==1.0.3
django-rest-swagger==0.3.2
django-rest-auth==0.5.0
djangorestframework==3.2.3
docopt==0.4.0
ecdsa==0.13
feedparser==5.2.1
gunicorn==19.3.0
kombu==3.0.26
langdetect==1.0.5
mandrill==1.0.57
Markdown==2.6.2
numpy==1.9.2
paramiko==1.15.2
psycopg2==2.6
pycrypto==2.6.1
pystache==0.5.4
pytz==2015.2
requests==2.6.2
six==1.9.0
static3==0.6.1
timestring==1.6.2
ads
